pyproject.toml
[project]
name = "propzy"
version = "0.1.0"
description = "Sistema de Gestão de Projetos da Ecoporé"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "python-decouple>=3.8",
    "psycopg2-binary>=2.9.9",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "redis>=5.0.1",
    "django-redis>=5.4.0",
    "gunicorn>=21.2.0",
    "django-allauth>=0.63.3",
    "django-crispy-forms>=2.4",
    "crispy-bootstrap5>=2025.6",
]

[project.optional-dependencies]
dev = ["ruff>=0.14.0", "mypy>=1.18.0", "django-stubs>=5.2.0", "django-stubs-ext>=5.2.0"]

s3 = ["boto3>=1.34.0", "django-storages>=1.14.2"]

[tool.setuptools.packages.find]
exclude = ["docker*", "locale*", "static*", "templates*", "staticfiles*", ".venv", ".vscode", ".github"]

# Configuração do Ruff
[tool.ruff]
# Mesma largura de linha do Black
line-length = 120
target-version = "py313"

# Diretórios a serem excluídos
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "staticfiles",
    "locale",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Regras habilitadas
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "DJ",  # flake8-django
]

# Regras ignoradas (ajuste conforme necessário)
ignore = [
    "E501",  # line too long (já controlado pelo line-length)
    "B008",  # do not perform function calls in argument defaults (comum em Django)
    "DJ001", # null=True em CharField/TextField (às vezes necessário)
]

# Permitir autofix para todas as regras habilitadas
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Ignorar imports não utilizados em __init__.py
"__init__.py" = ["F401"]
# Permitir todos os imports em settings.py
"settings.py" = ["F403", "F405"]
# Ignorar regras específicas em arquivos de teste
"test_*.py" = ["S101"]
"**/tests/**/*.py" = ["S101"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "apps.accounts", "apps.main", "config"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
# Usar aspas duplas
quote-style = "double"
# Usar trailing commas
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Configuração do Pylance/Pyright (para melhor integração)
[tool.pyright]
include = ["apps", "apps.accounts", "apps.main", "config"]
exclude = [
    "**/migrations",
    "**/__pycache__",
    "**/node_modules",
    ".venv",
    "staticfiles",
    "locale",
]
pythonVersion = "3.13"
pythonPlatform = "Linux"

# Configurações de tipo
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = false  # Ruff já faz isso
reportUnusedClass = false
reportUnusedFunction = false
reportDuplicateImport = false  # Ruff já faz isso
reportOptionalSubscript = false
reportOptionalMemberAccess = false
reportOptionalCall = false
reportOptionalIterable = false
reportOptionalContextManager = false
reportOptionalOperand = false
reportGeneralTypeIssues = false  # Django tem muitos padrões que confundem o Pylance
reportPrivateImportUsage = false
reportConstantRedefinition = false

# Configurações específicas do Django
[tool.django-stubs]
django_settings_module = "config.settings"

# Configuração do mypy para usar o plugin do django-stubs
[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.13"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true

# Configuração específica para apps Django
[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true
